+++ script.js
@@ -121,4 +121,31 @@ document.addEventListener("DOMContentLoaded", () => {
   }
 
   requestAnimationFrame(moveSharks)
+
+  // Spawn sharks on button click
+  document.querySelector("#shark-button").addEventListener("click", (event) => {
+    const dX = sharks.length % 2 === 0 ? 1 : -1
+    const dY = sharks.length % 2 === 0 ? 1 : -1
+
+    // Fallback to random value if mouse event has 0 x / y coordinate
+    // This can occur if click event is triggered by keyboard
+    const x = event.clientX || Math.floor(Math.random() * window.innerWidth)
+    const y = event.clientY || Math.floor(Math.random() * window.innerHeight)
+    const shark = {
+      x,
+      y,
+      dX: dX * (Math.floor(Math.random() * 20) + 20),
+      dY: dY * (Math.floor(Math.random() * 20) + 20),
+    }
+
+    const elem = document.createElement("div")
+    elem.className = "shark text-2xl"
+    elem.style.top = `${shark.y}px`
+    elem.style.left = `${shark.x}px`
+    elem.style.animationDuration = `${Math.floor((Math.random() * 5) + 1)}s`
+
+    shark.elem = elem
+    sharks.push(shark)
+    document.body.appendChild(elem)
+  })
 })